+++ classes/Service/PackageMarkdownService.php
@@ -247,11 +247,13 @@ class PackageMarkdownService
 
             flock($lock, LOCK_UN);
             fclose($lock);
+            @unlink($lockPath);
 
             return $backup;
         } catch (\Throwable $e) {
             flock($lock, LOCK_UN);
             fclose($lock);
+            @unlink($lockPath);
             throw $e;
         }
     }